Added Auth to Properties Endpoints

+++ properties/Properties/Controllers/PropertiesController.cs
@@ -5,6 +5,11 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using System;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Primitives;
+using System.Text;
+using System.Net;
 
 namespace Properties.Controllers
 {
@@ -18,53 +23,95 @@ namespace Properties.Controllers
         {
             _db = db;
         }
+        private bool Authenticate()
+        {
+            var request = this.Request.Headers;
+            StringValues token = " ";
+            if (request.TryGetValue("token", out token))
+            {
+                using (var client = new HttpClient())
+                {
+                    var req = new HttpRequestMessage(HttpMethod.Post, "http://localhost:5001/introspection/")
+                    {
+                        Content = new StringContent("{\"token\":\""+request["token"]+"\"}" , Encoding.UTF8, "application/json")
+                     };
+
+                    var res = client.SendAsync(req);
+                    
+                    var result = res.Result.StatusCode;
+
+                    if (result.ToString() == "OK")
+                    {
+                        return true;
+                    }
+                }
+
+            }
+            return false;
+        }
 
         [HttpGet]
-        public IEnumerable<Property> GetAll()
+        public IActionResult GetAll()
         {
-            return _db.Properties.Include(p => p.Schedules).ToList();
+            return Ok(_db.Properties.Include(p => p.Schedules).ToList());
         }
 
         [HttpGet("{id}")]
-        public Property GetById(int id)
+        public IActionResult GetById(int id)
         {
-            return _db.Properties.Find(id);
+            return Ok(_db.Properties.Find(id));
         }
 
         [HttpPost]
-        public void Create(Property property)
+        public IActionResult Create(Property property)
         {
-            _db.Properties.Add(property);
-            _db.SaveChanges();
+            if (Authenticate())
+            {
+                _db.Properties.Add(property);
+                _db.SaveChanges();
+                return Ok(property);
+
+            }
+            return Unauthorized();
         }
 
         [HttpPost("Delete/{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
-            Property property = _db.Properties.Find(id);
-            _db.Properties.Remove(property);
-            _db.SaveChanges();
+            if (Authenticate())
+            {
+                Property property = _db.Properties.Find(id);
+                _db.Properties.Remove(property);
+                _db.SaveChanges();
+                return Ok();
+            }
+            return Unauthorized();
         }
 
         [HttpPost("Update/")]
-        public void Update(Property property)
+        public IActionResult Update(Property property)
         {
-            _db.Properties.Update(property);
-            _db.SaveChanges();
+            if (Authenticate())
+            {
+                _db.Properties.Update(property);
+                _db.SaveChanges();
+                return Ok(property);
+            }
+            return Unauthorized();
         }
 
         [HttpGet("Schedule/{schedule_id}/{property_id}")]
-        public Schedule GetSchedule(int schedule_id, int property_id)
+        public IActionResult GetSchedule(int schedule_id, int property_id)
         {
             List<Property> properties = _db.Properties.Include(p => p.Schedules).Where(p => p.PropertyId == property_id).ToList();
             if (properties.Count() > 0)
             {
                 List<Schedule> schedules = properties[0].Schedules.Where(s => s.Id == schedule_id).ToList();
                 if (properties.Count() > 0) {
-                    return schedules[0];
+                    return Ok(schedules[0]);
                 }          
             }
-            throw new Exception("Schedule not found");
+            return NotFound();
         }
     }
 }
